Change method of removing array element
Some notes to myself:
I discovered classes and the static method during
this project. This was better than using just an
object constructor, because without static I was
using a conditional statement to initialize a
satic variable. Nothing wrong with this, but I
think it's ever so slightly easier to read now.
This was alongside the other benefits of using
classes that I read about.

Also, for this project, Odin recommended as an easy
giving the dynamically generated DOM elements
a data-attribute that corresponds to the index of
the library itself. This is in my opinion not
actually a very easy or good solution. When an
item is removed from the array, it shifts, so the
DOM elements must be constantly updated. Instead
I gave each instantiated object a unique ID /not
a very robust one/ and used